Credits Work, Death Cutscene Implementation Change, & Intro Cutscene Saving Fix
We've adjusted the credits slightly, adding a manager for the death cutscene and having an option for the end credits which doesn't trigger the 'Game Over' screen, but instead just brings you back to the main menu.

We've also changed the 'FootstepClipContainer' to a more standard 'AudioClipContainer', and are using this with the Death Cutscene for the death and bite SFX.

Additionally, we've fixed a bug with the Intro Cutscene Camera which was causing it to break the hub partitioning when loading from a save into a different area.