Credits End Sequence, Activation Trigger, & Singleton Debug Fixes
We've created the Credits End Sequence, along with finalising some bits of the Credits Sequence in general.

Additionally, we've created an 'Activation Trigger' which calls when the object is first enabled, along with fixing the debug log error for the Singleton<T> and ProtectedSingleton<T> generic classes so that they properly display their type, rather than always displaying "SceneLoader" as their type.

Furthermore, we've also made it so that the EndCreditsUI resets the audio levels AFTER we finish loading the MainMenuScene to prevent a sudden short spike of audio slipping through.