Tutorial Text Display & VolumeTrigger OnExit option
We've worked on the Tutorial Text Display, getting it to the point we desire for our implementation. We can not toggle it on & off, and there is an option for flickering text.

Additionally, we've added the ability to have VolumeTriggers which only activate when the desired entity exits the trigger, rather than entering it.